+++ CHANGELOG.md
@@ -9,7 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-* Nothing, yet!
+### Added
+
+* Open Telemetry and Jaeger tracing support
 
 ## [0.14.0] - 04/16/2022